requirements.txt
fastapi
uvicorn
numpy
Pillow
keras